+++ lib/model/settings/enums/video_auto_play_mode.dart
@@ -3,14 +3,14 @@ import 'package:flutter/widgets.dart';
 
 import 'enums.dart';
 
-extension ExtraSlideshowVideoPlayback on SlideshowVideoPlayback {
+extension ExtraSlideshowVideoPlayback on VideoAutoPlayMode {
   String getName(BuildContext context) {
     switch (this) {
-      case SlideshowVideoPlayback.skip:
-        return context.l10n.videoPlaybackSkip;
-      case SlideshowVideoPlayback.playMuted:
+      case VideoAutoPlayMode.disabled:
+        return context.l10n.settingsDisabled;
+      case VideoAutoPlayMode.playMuted:
         return context.l10n.videoPlaybackMuted;
-      case SlideshowVideoPlayback.playWithSound:
+      case VideoAutoPlayMode.playWithSound:
         return context.l10n.videoPlaybackWithSound;
     }
   }